first version of pubnubwrapper working
